Created snapshot functionalit, start/stop btn, configuration snapshot
Snapshot functionality added to dashboard and configuration page, every 60s a new frame is pulled from rabbitMQ (if available) and displayed instead of an offline image when a feed is active. 
Added start / stop button to feeds to start feeds from the front end, added offline image when feeds are inactive/ not sending frames.
Added snapshot image to configuration screen as preview.